requirements.txt
# python packages to install, one per line
transformers==4.40.2
diffusers==0.28.0
ipywidgets==8.1.5
accelerate==1.0.1
huggingface_hub[cli,torch]==0.21.4
jupyterlab>3.0
notebook==6.4.10
torchvision==0.16.0
nvtx==0.2.5
scipy==1.14.1
dataclass_wizard==0.23.0
gradio==4.35.0
jinja2==3.1.2
uvicorn==0.32.0
peft==0.9.0
fastapi==0.111.0
